Move functions to v1-nxm-push-utils
